LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/postmessage-from-waiting-serviceworker.https-expected.txt



FAIL Client.postMessage() from waiting serviceworker. assert_unreached: unexpected rejection: assert_equals: should be same registration expected object "[object ServiceWorkerRegistration]" but got object "[object ServiceWorkerRegistration]" Reached unreachable code
